SEMECO-review-v1: No body for 4XX responses
A 400, 401 or 403 response is returned when
something is wrong with the request in general
(has no body, body is not formatted correctly,
no Authorization header was added to the request,
Authorization header contains invalid credentials,
etc.) As such, no valid response can (and should)
be returned and therefore the response will be one
with a status code in the 4XX range and no body.

In case something is wrong with the contents of
the request, one should fall back to 400 if the
server cannot make heads or tails of it, but
otherwize a 200 status code should be returned
and the body of the response should contain any
feedback on the request body. So if a batch of
events is sent and one of the events cannot be
processed, the response should be 200 with a body
containing details of the failure. Remember that
the rest of the events will still be processed,
so only the failed event should be resent.